photovault_django/requirements.txt
# Django Core
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-environ>=0.11.0

# Database
psycopg2-binary>=2.9.0
pgvector>=0.2.0

# Authentication & Security
djangorestframework-simplejwt>=5.3.0
argon2-cffi>=23.1.0
cryptography>=41.0.0
django-ratelimit>=4.1.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0,<2.0.0
face-recognition>=1.3.0
dlib>=19.24.0

# Background Tasks
celery[redis]>=5.3.0
redis>=5.0.0

# API Documentation
drf-spectacular>=0.26.0

# Utilities
python-slugify>=8.0.0
qrcode[pil]>=7.4.2
python-dotenv>=1.0.0
requests>=2.31.0
python-json-logger>=2.0.0

# Production Server
gunicorn>=21.2.0
whitenoise>=6.6.0

# Development & Quality Tools
ruff>=0.1.6
black>=23.11.0
mypy>=1.7.0
pytest>=7.4.0
pytest-django>=4.7.0
pytest-cov>=4.1.0
bandit>=1.7.5
pip-audit>=2.6.0
pre-commit>=3.5.0